serving/requirements.txt
streamlit
plotly
pandas
kafka-python
pydeck
pydeck